refactor: split main app into smaller functions
