+++ streamlit_app.py
@@ -1,37 +1,11 @@
 import streamlit as st
-import json
 import time
 
-from helpers.drive.api import GSAuthentication, download_datasets
-from helpers.utils.dataset import generate_incomplete_dataset, get_df_incomplete, save_results, find_uncertain_y_indexes
+from helpers.streamlit.callbacks import update_point_color
+from helpers.streamlit.utils import init_session_state, read_settings_from_file, validate, init_new_annotation_task
+from helpers.utils.dataset import add_projection_dimensions_to_df_incomplete
 from helpers.utils.projections import projections_dict, plot_scatter
-from sklearn.cluster import KMeans
 
-
-def validate(user_id):
-    validated = user_id.isdigit()
-    if validated:
-        st.session_state['id'] = user_id
-    return validated
-
-
-def calculate_y_pred_uncertain(df_incomplete):
-    X = df_incomplete[reference_columns].values
-
-    kmeans = KMeans(n_clusters=number_of_clusters, random_state=0)
-    kmeans.fit(X)
-    y_pred = kmeans.predict(X)
-
-    uncertain_y_indexes = find_uncertain_y_indexes(X, number_of_clusters, fuzzy_certainty_thres=0.6)
-    incomplete_indexes = df_incomplete[df_incomplete[incomplete_column].isnull()].index.tolist()
-    uncertain_incomplete_y_indexes = set(uncertain_y_indexes).intersection(incomplete_indexes)
-    y_pred_uncertain = [int(el) if i not in uncertain_incomplete_y_indexes else None for i, el in enumerate(y_pred)]
-    return X, y_pred_uncertain
-
-
-def update_session_state(X, y_pred_uncertain):
-    st.session_state.X = X
-    st.session_state['y_pred_uncertain'] = y_pred_uncertain
 
 
 def get_y_pred_uncertain_str():
@@ -44,64 +18,31 @@ user_id = st.sidebar.text_input('Your ID')
 user_password = st.sidebar.text_input('Password', type="password")
 
 
-def update_point_color():
-    df_incomplete.at[st.session_state['current_null_index'], incomplete_column] = st.session_state['my_slider']
-
-
 if not validate(user_id) or not user_password == st.secrets['password']:
     '# Please authorize'
     'Please provide your ID and password in the sidebar inputs.'
 else:
-    with open("settings.json", "r") as f:
-        settings = json.load(f)
-
-    datasets_settings = settings['data']
-
-    if 'gsheet' not in st.session_state:
-        st.session_state["gsheet"] = GSAuthentication()
-    if 'data' not in st.session_state:
-        with st.spinner(text="In progress..."):
-            download_datasets(st.session_state["gsheet"], [dataset['name'] for dataset in datasets_settings])
-            st.session_state['data'] = True
-    if 'started' not in st.session_state:
-        st.session_state['started'] = False
-    if 'annotated_points' not in st.session_state:
-        st.session_state['annotated_points'] = 0
+    datasets_settings = read_settings_from_file()
+    init_session_state(datasets_settings)
 
     with st.sidebar:
         st.markdown('## Settings')
         dataset_name_selectbox = st.selectbox('Select dataset', [dataset['name'] for dataset in datasets_settings])
-        na_fraction_selectbox = st.selectbox('Select na_fraction', [0.05, 0.1, 0.15, 0.2, 0.25, 0.3])
+        na_fraction_selectbox = st.selectbox('Select na_fraction', [0.05, 0.1, 0.15, 0.2, 0.25, 0.3], key="na_fraction_selectbox")
 
     dataset_settings = next(dataset for dataset in datasets_settings if dataset['name'] == dataset_name_selectbox)
-
-    number_of_clusters = dataset_settings['number_of_clusters']
-    reference_columns = dataset_settings['reference_columns']
     incomplete_column = dataset_settings['incomplete_column']
-    target_column = dataset_settings['target_column']
 
     if st.sidebar.button('START'):
-        if 'dataset_generation_seed' not in st.session_state:
-            st.session_state['dataset_generation_seed'] = int(time.time())
-
-        df_incomplete = generate_incomplete_dataset(st.session_state['dataset_generation_seed'],
-                                                    dataset_settings['name'],
-                                                    incomplete_column,
-                                                    na_fraction_selectbox)
-        update_session_state(*calculate_y_pred_uncertain(df_incomplete))
-        st.session_state['annotator_error'] = 0
-        st.session_state['mean_error'] = 0
-        if None not in st.session_state['y_pred_uncertain']:
-            'Sorry, there are no border points to annotate. Choose other dataset or na_fraction.'
-        else:
-            st.session_state['started'] = True
+        st.session_state.df_incomplete = init_new_annotation_task(dataset_settings)
 
     if st.session_state['started']:
         with st.sidebar:
             projection_key = st.selectbox('Set projection', projections_dict.keys())
         try:
-            df_incomplete = get_df_incomplete(str(projection_key), st.session_state.X, dataset_settings,
-                                              na_fraction_selectbox)
+            df_incomplete = add_projection_dimensions_to_df_incomplete(str(projection_key),
+                                                                       st.session_state.X,
+                                                                       st.session_state.df_incomplete)
 
             df_incomplete['y_pred'] = get_y_pred_uncertain_str()
 
@@ -119,7 +60,7 @@ else:
                 plot_scatter(df_to_show, incomplete_column, current_null_index)
 
                 value = st.slider("", float(df_to_show[incomplete_column].min()), float(df_to_show[incomplete_column].max()),
-                                  step=0.1, key='my_slider',  on_change=update_point_color)
+                                  step=0.1, key='my_slider',  on_change=update_point_color, args=[df_incomplete, incomplete_column])
 
                 f"Records to be labeled: {len(df_null_indexes)}"
                 st.progress(
